+++ packages/radonis-shared/src/index.ts
@@ -11,4 +11,6 @@ export { isClient, isProduction, isServer } from './environment'
 export { Exception } from './exception'
 export { definePlugin, PluginsManager } from './pluginsManager'
 export type { Plugin } from './types'
-export { interpolate, separateArray, stringifyAttributes } from './utils'
+export { interpolate } from './utils/interpolate'
+export { separateArray } from './utils/separateArray'
+export { stringifyAttributes } from './utils/stringifyAttributes'
